backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.2
pydantic-settings>=2.6.1
supabase==2.22.4
realtime==2.22.4
openai>=1.51.0
xgboost>=2.1.1
scikit-learn>=1.6.0
pandas>=2.2.3
numpy>=2.0.0
python-multipart>=0.0.12
python-dotenv>=1.0.1
httpx>=0.27.2
sse-starlette>=2.1.3
joblib>=1.4.2
openpyxl>=3.1.5
